+++ app/_components/products-section.tsx
@@ -36,7 +36,7 @@ const ProductsSection = () => {
       <div className="max-w-[1200px] mx-auto flex flex-col items-center justify-center gap-6 px-4">
         <Title text="Products" borderColor="[#1D9B5E]/40" textColor="#1D9B5E" />
         <div className="max-w-[950px]">
-          <h2 className="font-semibold text-[28px] text-[#020617]">
+          <h2 className="font-semibold text-lg lg:text-[28px] text-[#020617]">
             Discover financial products designed to help you save, grow, and
             achieve more — with convenience, security, and trusted support every
             step.
