+++ app/_components/testimonials.tsx
@@ -47,14 +47,14 @@ const testimonialsData = [
 
 const Testimonials = () => {
   return (
-    <section className="py-6 lg:py-16 flex flex-col items-center justify-center gap-6 px-4">
+    <section className="py-6 max-w-[1400px] mx-auto lg:py-16 flex flex-col items-center justify-center gap-6 px-4">
       <Title
         text="Testimonials"
         borderColor="[#1D9B5E]/40"
         textColor="#1D9B5E"
       />
       <div className="max-w-[945px] mx-auto mb-8">
-        <h2 className="font-semibold text-[28px] text-[#020617]">
+        <h2 className="font-semibold text-xl lg:text-[28px] text-[#020617]">
           Hear what our customers are saying about their experience — real
           stories of trust, growth, and satisfaction with Covenant MFB.
         </h2>
