+++ app/about-us/_components/team-data.tsx
@@ -1,0 +1,33 @@
+export type TeamMember = {
+  name: string;
+  role: string;
+  imageUrl: string;
+  id:number
+};
+
+export const boardMembers: TeamMember[] = [
+  {id:1, name: 'Adrs. Bola, Shipe Olaseni FCIA', role: 'Board President', imageUrl: '/Group1.png' },
+  { id:2,name: 'Dr. Barry Roberts ORM, FICA', role: 'MD/CEO/Board MD', imageUrl: '/Group2.png' },
+  { id:3,name: 'Prof. Uwalomwa Uwuigbe', role: 'Chairman, Board Audit Committee', imageUrl: '/Group3.png' },
+  { id:4,name: 'Prof Chinonye Love Moses', role: 'Chairman, T-Finance/T-Credit', imageUrl: '/Group4.png' },
+  { id:5,name: 'Prof. Uwalomwa Uwuigbe', role: 'Chairman, Board Audit Committee', imageUrl: '/Group3.png' },
+  { id:6,name: 'Prof Chinonye Love Moses', role: 'Chairman, T-Finance/T-Credit', imageUrl: '/Group4.png' },
+];
+
+export const managementStaff: TeamMember[] = [
+  {id:1, name: 'Bekain Oliseya', role: 'Head, Admin and Corp Services', imageUrl: '/Group1.png' },
+  { id:2,name: 'Oluwaseun Tunji-Salawu', role: 'Branch Manager, Ota Ogun State', imageUrl: '/Group2.png' },
+  { id:3,name: 'Koseriphr Michael Oludipe', role: 'Branch Manager, Akungba, Ondo', imageUrl: '/Group3.png' },
+  { id:4,name: 'Bekain Oliseya', role: 'Head, Admin and Corp Services', imageUrl: '/Group1.png' },
+  { id:5,name: 'Oluwaseun Tunji-Salawu', role: 'Branch Manager, Ota Ogun State', imageUrl: '/Group2.png' },
+  { id:6,name: 'Koseriphr Michael Oludipe', role: 'Branch Manager, Akungba, Ondo', imageUrl: '/Group3.png' },
+  {id:7, name: 'Bekain Oliseya', role: 'Head, Admin and Corp Services', imageUrl: '/Group1.png' },
+  { id:8,name: 'Oluwaseun Tunji-Salawu', role: 'Branch Manager, Ota Ogun State', imageUrl: '/Group2.png' },
+  { id:9,name: 'Koseriphr Michael Oludipe', role: 'Branch Manager, Akungba, Ondo', imageUrl: '/Group3.png' },
+  { id:10, name: 'Bekain Oliseya', role: 'Head, Admin and Corp Services', imageUrl: '/Group1.png' },
+  {id:11, name: 'Oluwaseun Tunji-Salawu', role: 'Branch Manager, Ota Ogun State', imageUrl: '/Group2.png' },
+  {id:12, name: 'Koseriphr Michael Oludipe', role: 'Branch Manager, Akungba, Ondo', imageUrl: '/Group3.png' },
+];
+
+
+
