+++ app/about-us/page.tsx
@@ -1,31 +1,25 @@
 import NumbersSection from "../_components/numbers-section";
 import Testimonials from "../_components/testimonials";
-import AboutHeroSection from "./about-us";
-import OurHistorySection from "./our-history";
-import WhoWeAreSection from "./who-we-are";
+import AboutHeroSection from "./_components/about-us";
+import OurHistorySection from "./_components/our-history";
+import WhoWeAreSection from "./_components/who-we-are";
 import Carousel from "../_components/carousel";
-import MeetTheTeamSection from "./meet-the-team";
-import TimelineSection from "./timeline-section";
+import MeetTheTeamSection from "./_components/meet-the-team";
+import TimelineSection from "./_components/timeline-section";
 
 const AboutPage = () => {
   return (
     <div>
-  <AboutHeroSection/>
- <WhoWeAreSection/>
- <NumbersSection/>
- <OurHistorySection/>
- <TimelineSection/>
- <MeetTheTeamSection/>
-<Testimonials/>
-<Carousel/>
-
- 
-  </div>
-  
-  )
-  
+      <AboutHeroSection />
+      <WhoWeAreSection />
+      <NumbersSection />
+      <OurHistorySection />
+      <TimelineSection />
+      <MeetTheTeamSection />
+      <Testimonials />
+      <Carousel />
+    </div>
+  );
 };
 
 export default AboutPage;
-
-
